+++ aprofplot/src/aprofplot/SmsEntry.java
@@ -15,8 +15,8 @@ public class SmsEntry implements Comparable<SmsEntry> {
     private long min_cost;
     private long max_cost;
     private long cost_sum;
-    private long cost_sqr_sum;
-    private int occ;
+    private double cost_sqr_sum;
+    private long occ;
 
     private double ratio;
 
@@ -26,7 +26,7 @@ public class SmsEntry implements Comparable<SmsEntry> {
     private static double[] ratio_config = {1, 0, 0};
 
     public SmsEntry(int sms, long min_cost,
-            long max_cost, long cost_sum, long cost_sqr_sum, int occ) {
+            long max_cost, long cost_sum, double cost_sqr_sum, long occ) {
 
         this.sms = sms;
         this.min_cost = min_cost;
@@ -104,7 +104,7 @@ public class SmsEntry implements Comparable<SmsEntry> {
         return 0;
     }
 
-    public int getOcc() {
+    public long getOcc() {
         return occ;
     }
